refactor: clean namespaces
